+++ endpoints/instantiations/api_key.py
@@ -2,7 +2,7 @@ from flask import Response
 
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.instantiations import STANDARD_JWT_AUTH_INFO
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.primary_resource_logic.api_key import create_api_key_for_user
 
 from endpoints.schema_config.enums import SchemaConfigs
