+++ endpoints/instantiations/archives.py
@@ -9,7 +9,7 @@ from middleware.security.auth.info.instantiations import (
     ARCHIVE_WRITE_AUTH_INFO,
     API_OR_JWT_AUTH_INFO,
 )
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.primary_resource_logic.archives_queries import (
     archives_get_query,
     update_archives_data,
