+++ endpoints/instantiations/batch.py
@@ -8,7 +8,7 @@ from endpoints.schema_config.instantiations.bulk.data_sources import (
 )
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.instantiations import STANDARD_JWT_AUTH_INFO
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.primary_resource_logic.bulk import (
     bulk_post_agencies,
     bulk_post_data_sources,
