+++ endpoints/instantiations/github_data_requests.py
@@ -1,6 +1,6 @@
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.base import AuthenticationInfo
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.enums import AccessTypeEnum, PermissionsEnum
 from middleware.primary_resource_logic.github_issue_app import (
     synchronize_github_issues_with_data_requests,
