+++ endpoints/instantiations/map.py
@@ -3,7 +3,7 @@ from flask import Response
 from config import limiter
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.instantiations import API_OR_JWT_AUTH_INFO
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.primary_resource_logic.data_sources import (
     get_data_sources_for_map_wrapper,
 )
