+++ endpoints/instantiations/metadata.py
@@ -2,7 +2,7 @@ from flask import Response
 
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.instantiations import API_OR_JWT_AUTH_INFO
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.primary_resource_logic.metadata import (
     get_record_types_and_categories,
 )
