+++ endpoints/instantiations/oauth.py
@@ -5,7 +5,7 @@ from endpoints.schema_config.instantiations.auth.github.oauth import (
 )
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.instantiations import NO_AUTH_INFO
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.enums import CallbackFunctionsEnum
 from middleware.schema_and_dto.dtos.github.oauth import GithubOAuthRequestDTO
 from middleware.third_party_interaction_logic.callback.flask_sessions import (
