+++ endpoints/instantiations/refresh_session.py
@@ -2,7 +2,7 @@ from flask import Response
 
 from middleware.security.access_info.refresh import RefreshAccessInfo
 from middleware.security.auth.info.base import AuthenticationInfo
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.enums import AccessTypeEnum
 from middleware.primary_resource_logic.login_queries import (
     refresh_session,
