+++ endpoints/instantiations/request_reset_password.py
@@ -5,7 +5,7 @@ from endpoints.schema_config.instantiations.reset_password.request import (
 )
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.instantiations import NO_AUTH_INFO
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.primary_resource_logic.reset_token_queries import request_reset_password
 from endpoints.schema_config.enums import SchemaConfigs
 from endpoints._helpers.response_info import ResponseInfo
