+++ endpoints/instantiations/search/follow/national.py
@@ -7,7 +7,7 @@ from endpoints.schema_config.instantiations.search.follow.national import (
 )
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.instantiations import STANDARD_JWT_AUTH_INFO
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.primary_resource_logic.search.wrappers.follow.national.follow import (
     follow_national_wrapper,
 )
