+++ endpoints/instantiations/search/location_and_record_type.py
@@ -9,7 +9,7 @@ from endpoints.schema_config.instantiations.search.location_and_record_type impo
 )
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.instantiations import API_OR_JWT_AUTH_INFO
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.primary_resource_logic.search.wrappers.record_type_and_record_categories import (
     search_wrapper,
 )
