+++ endpoints/instantiations/typeahead_suggestions.py
@@ -6,7 +6,7 @@ from config import limiter
 from db.client.core import DatabaseClient
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.instantiations import NO_AUTH_INFO
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.primary_resource_logic.typeahead_suggestion import (
     get_typeahead_results,
 )
