+++ endpoints/instantiations/user_profile.py
@@ -3,7 +3,7 @@ from flask import Response
 from endpoints.schema_config.instantiations.user.put import UserPutEndpointSchemaConfig
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.security.auth.info.instantiations import STANDARD_JWT_AUTH_INFO
-from middleware.decorators.decorators import endpoint_info
+from middleware.decorators.endpoint_info import endpoint_info
 from middleware.primary_resource_logic.reset_token_queries import (
     change_password_wrapper,
 )
